Refactor FileTransferAgent facade classes into base class
